backend/requirements.txt
Django==4.2.5
psycopg2-binary
gunicorn
djangorestframework
djangorestframework-simplejwt
alibabacloud_dysmsapi20170525
alibabacloud_tea_openapi
alibabacloud_tea_util
python-dotenv
django-redis
pillow
numpy<2
urllib3==1.26.13
httpx==0.27.2
openai==1.7.0
pytest
pytest-cov
coverage
mutagen